backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn==0.24.0
better-exceptions==0.3.2
asyncpg==0.29.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0